Move amp Article component to pages
